Move static db helper reference
