+++ collect_app/src/main/java/org/odk/collect/android/database/FormsDatabaseHelper.java
@@ -31,12 +31,35 @@ import static org.odk.collect.android.database.DatabaseConstants.FORMS_DATABASE_
  * This class helps open, create, and upgrade the database file.
  */
 public class FormsDatabaseHelper extends SQLiteOpenHelper {
-    private final DatabaseMigrator databaseMigrator;
+
+    private static FormsDatabaseHelper dbHelper;
+
+    public static synchronized FormsDatabaseHelper getDbHelper() {
+        if (dbHelper == null) {
+            recreateDatabaseHelper();
+        }
+
+        return dbHelper;
+    }
 
     public FormsDatabaseHelper(DatabaseMigrator databaseMigrator, StoragePathProvider storagePathProvider) {
         super(new DatabaseContext(storagePathProvider.getOdkDirPath(StorageSubdirectory.METADATA)), FORMS_DATABASE_NAME, null, FORMS_DATABASE_VERSION);
         this.databaseMigrator = databaseMigrator;
     }
+
+    public static void recreateDatabaseHelper() {
+        dbHelper = new FormsDatabaseHelper(new FormDatabaseMigrator(), new StoragePathProvider());
+    }
+
+    @SuppressWarnings("PMD.NonThreadSafeSingleton") // PMD thinks the `= null` is setting a singleton here
+    public static void releaseDatabaseHelper() {
+        if (dbHelper != null) {
+            dbHelper.close();
+            dbHelper = null;
+        }
+    }
+
+    private final DatabaseMigrator databaseMigrator;
 
     @Override
     public void onCreate(SQLiteDatabase db) {
